NbtkTable: Take care to avoid accessing uninitialised memory
When calculating the column width it's possible for the table to access
unallocated memory if the column/row position plus the row/column-span is
greater than the number of columns/rows in the table. This change fixes that
by checking the that the array index does not exceed the bounds.