Port TextureFrame to the new materials and Clone API
COGL now uses the "materials" concept, which allows to decouple
the path primitives from the fill operations. It is possible
to just draw something and then fill it with a texture handle.

TextureFrame should update to the new API.

Also, the CloneTexture actor has been removed from the public
API, and superceded by the Clone actor.

Since we're interested in cloning textures, we can simply
subclass ClutterActor and provide the methods and properties
that CloneTexture used.