[build] Update the autofoo in NBTK
NBTK is currently missing:

  - libtool flags
  - a sane sources listing
  - versioning macros
  - debug macros
  - maintainer flags

The private headers are also installed, and the enumeration types generator
is scanning headers that it should not scan.

This commit makes the automake usage more reliable and less error-prone; it
also updates the configure template.

Part of the whole autofoo is adapted from Clutter's own.