[build] Fix a compiler warning
The ccss_style_get_property() function uses a void** to store the
value of a property. Using a ccss_property_t* breaks the aliasing
rules and gcc is very angry about that.

Instead of using a ccss_property_t* we just use a gpointer - which
is a void* - and then do the assignment:

  ccss_property_t *border_image = value;

Only if the ccss_style_get_property() was successful.