[build] Version header files and shared object
Instead of installing the NBTK header files into:

  ${includedir}/nbtk

And installing an anonymous:

  ${libdir}/libnbtk.so

Shared object we should version both header files and library
to allow:

  - clean include paths
  - parallel installation

Like most other pkg-config based libraries.

The headers should be installed under:

  ${includedir}/nbtk-${nbtk_api_version}/nbtk

And the shared object should be:

  ${libdir}/libnbtk-${nbtk_api_version}.so

Bonus points for making it all depend on a single macro defined
in the configure.ac template file.