Add NbtkBin, a container for one child
NbtkWidget is an abstract class that not only implements
a stylable element of the scenegraph, but also a container
for a single child.

NbtkWidget should just be the base class, and deal with styling,
since subclasses of NbtkWidget might not want to be containers.

To provide a replacement for the single-child-container feature,
NbtkBin has been added.

NbtkBin is a simple container, able only to keep a single child;
it is style through the NbtkWidget stylable properties.