[scroll-view] Make ScrollView inherit from Bin
A ScrollView is a stylable container with only one child.

ScrollView imposes restrictions on the Container::add() implementation
so we need to override the Container interface; we don't want to reimplement
the whole container machinery, otherwise we wouldn't be inheriting from
NbtkBin in the first place. Since the limitations are *on top* then we
just need to get a pointer to the parent class implementation using:

  g_type_interface_peek_parent()

And then chain up.