Remove DepthGroup
The DepthGroup is used only by the scrollable code test cases, since
they require a container with depth testing to avoid incorrect paints
of the child rectangles.

The depth testing can be successfully enabled using the ::paint signal
which was not available at the time DepthGroup was written in Tidy.