Add NbtkDraggable
NbtkDraggable is an interface for implementing actors and widgets
that can be dragged using the pointer device.

NbtkDraggable is an interface, so it can be added to existing
actors, but it also provide a default implementation of the dragging
machinery without requiring the implementation of the signal
handling.

In order to implement a draggable actor you should:

  - implement the NbtkDraggable signals drag-begin, drag-motion
    and drag-end
  - override the interface properties
  - call nbtk_draggable_enable() on your actor

It is also possible to override the implementation, or chain up
to it using the usual GInterface API.

An example of a draggable actor is provided.