Implement NbtkDroppable
The Droppable interface provides the drop target capabilities to an
actor.

In order to implement the NbtkDroppable interface, an Actor should
override the handlers for the ::over-in, ::over-out and ::drop
signals.

Conditional drop is implemented by overriding the ::accept_drop()
virtual function.

The entire machinery can be replaced by overriding the ::enable()
and ::disable() virtual functions.

Like NbtkDraggable, NbtkDroppable provides a default implementation
that makes implementing drop targets easy.