[draggable] Use ::captured-event instead of bubbled signals
The current dragging machinery is implemented through bubble-phase
signals for both motion and release. This prevents having the pointer
device from ever leaving the dragged actor, which is usually not what
you want.

Instead of using grabs we can use the capture-phase event handling
by simply attaching a handler to the ::captured-event signal on the
stage that contains a draggable actor.

The dragging must still be initiated by a ::button-press-event on
the draggable actor, but then it's all deferred to the ::captured-event
handler.