[droppable] Support multiple drop targets
The current Droppable implementation is limited to a single drop
target; as soon as new Droppable are added to the scene graph we
get replications due to the multiple ::captured-event connections.

Instead of connecting multiple times on Droppable::enable we need
to connect just once and keep all the enabled Droppables inside
the DropContext.

The logic of the over-in/-out detection is also pericularly flakey
and needs to be fixed.