[bin] Use the constrained size when allocating
The child actor of NbtkBin should be constrained to the
available size when computing its preferred size. This means
that calling get_preferred_size() is out of the question.

Instead, we should check the preferred geometry requisition,
get the natural size and clamp it between the minimum and
the available size.