[NbtkIconView] Add freeze/thaw API for bulk updates
If a user of NbtkIconView knows that there are going to be a large number of
changes to the model in short succession it can improve performance by
freezing the view with nbtk_icon_view_freeze (). This will prevent the view
from updating with the model until nbtk_icon_view_thaw () is called. When
thawing this will also cause the view to be updated.