[NbtkGtkLightSwitch] Update slider position on "configure-event"
When the state was set before the widget was anchored the logical
state ("active") and the appearance (slider position) went out of sync.
Fix this by recalculating the slider position in "configure-event".

Also update gtk-test so it can be run with an initial button state,
which is passed when launching (./gtk-test 1 / ./gtk-test 0).