[tests] Add "test-performance" for benchmarking.
The program gets a list of NbtkWidget subclasses from nbtk.types. Each
type is subsequently instantiated in an idle callback and put onto the
stage.

The benchmark is conducted by calling the "style-changed" handler on
each widgets. By default the handler is run 50 times, but an arbitrary
number can be passed on the commandline, e.g. "./test-performance 100".