[NbtkTextureCache] Remove the AUTO_MIPMAP flag and don't set the filters
Clutter now stores the texture filter quality in the material of the
ClutterTexture rather than in the cogl texture so there is no need to
set the filters. Instead each ClutterTexture can use its own filter
settings for the same cogl texture. There is also no need to set the
AUTO_MIPMAP flag because it is on by default and the mipmaps will only
be generated the first time the texture is rendered in high quality
mode.