[NbtkTextureFrame] Reference the parent texture with g_object_ref_sink
The floating reference is now taken on the parent texture when it is
set. This is needed to avoid a leak if nothing else takes a reference
to the texture. In particular, NbtkWidget creates a new ClutterTexture
to set for the background image and directly passes it to
nbtk_texture_frame_new without taking a reference. That caused it to
leak the ClutterTexture every time the style changed because nothing
unref'd the floating ref.