Create a subtexture and atlas-texture cache for nbtk
this patch creates a NbtkSubtexture class, which takes a bigger texture
as input, but which acts as if it's only a part of this bigger (shared) texture.

With this we can create atlas textures, or megatextures, which contain, say
all the icons in a directory in one texture. The icons are then still individual
Clutter actors, but they share the texture space and only need to load the texture
from disk into the GPU once.

The "makecache" program is used to turn a set of icons into a texture cache.