Move the "style-changed" signal up to the Stylable interface
This will allow the Stylable to be notified when one of the properties
changes, hence allowing NbtkStyle to know if it needs to create a new
query for the Styleable or not.

This change also changes the stylable-changed signal to a run first,
signal allowing the stylable implementation to block the changed signal
if subclasses should not update their style information (for example, if
the widget is not yet mapped).