[NbtkButton] Plug leak of priv->old_bg
Before its style changes, NbtkButton will take a reference to the old
background actor and store it in priv->old_bg. If it starts a
transition on that background it then also parents the actor. The
completion function for the animation unparents the actor and sets the
pointer to NULL but does not remove the extra ref. The dispose
function removes the ref but does not unparent it. That means there
will always be a leaked reference to the actor.

This can sometimes cause a critical error if the button is destroyed
before the transition animation is completed (such as if the button is
used as a close button). The button will be finalized but the
background would not be unparented and it will still be animated. It
will then get upset because it still has a pointer to an invalid
parent actor.

This patch tries to fix it by keeping track of whether we have
parented old_bg or not. There is a separate function for disposing
old_bg which unrefs the actor and unparents it if needed.

It now also avoids painting or allocating old_bg if the button did not
parent it - otherwise it will be allocated and painted twice in
NbtkFadeButton. It also adds the missing map and unmap calls for
old_bg.