[NbtkWidget] Add ::draw_background caller
Currently, NbtkWidget subclasses that override the ::paint virtual
function of ClutterActor need to chain up in order to get the
background image or color painted. Unfortunately, not every subclass
can do this: some times a widget might need to paint something between
the background and its children (e.g. a selection box).

If a class cannot chain up but still wishes to get the background
painted then it should be able to call a NbtkWidget method to get
the ::draw_backgound() virtual function invoked with the same values
as used by the default implementation of NbtkWidget::paint().

Fixes bug 4445

  http://bugzilla.moblin.org/show_bug.cgi?id=4445