[NbtkExpander] Hide an unexpanded child on add not paint
Previously the NbtkExpander would try to hide its child during the
paint implementation if the expander was not expanded. This causes
problems because hiding the actor would cause a redraw but as we are
already in a paint cycle the redraw_pending flag on the stage will get
lost. When something else tries to a queue another redraw later it
will short-circuit because the actor flags say the redraw is already
pending. This was causing moblin bug #3662.

Instead of hiding the child on paint it now overrides the
ClutterContainer::paint method and hides there.