Generate introspection data at build time
Update the build to generate the GObject Introspection data from
the source files and their comments. The generated data is in two
formats: the GIR (and XML document describing the API, types,
properties and signals); and the typelib (a compiled binary blob
that can be mmap()'ed by the GIRepository API).

The introspection data is generated for both NBTK and NBTK-GTK.