[draggable] Some fixes for thresholding + stand-in
If we use a stand-in actor, we can't transform coordinates based on that
until we actually emit the drag-begin signal (as it may not be parented
to the stage before then), so use the draggable until it passes the drag
threshold and we've emitted the drag-begin signal.

Also fix that it was possible to receive a drag-end signal without a
drag-begin if you dragged and released within the drag threshold.

Add a check in the stage-capture event handler to see if the button is
held down during the motion event - if it isn't, use the released
handler instead of the motion handler. This is necessary in case we miss
the release event, as in the case of grabs.