[NbtkEntry] Don't allow the entry to get key focus and don't pass on key events
Since commit 5f8e83b9 it is expected that the inner ClutterText actor
should get key focus and the NbtkEntry will only handle events that
aren't handled by the ClutterText and bubble up. However the
ClutterText actor can be considered an internal implementation detail
so it is natural for applications to try and set the key focus on the
actual NbtkEntry. Previously if this happened the entry would get
focus and it would be able to pass events down to the ClutterText but
clutter_text_focus_in_cb would not be called so the caret would not be
visible.

Instead, the NbtkEntry now forces the focus on to the ClutterText
whenever it gets focus. Now that the entry assumes the ClutterText has
focus there is no point in passing the unhandled events back
down. This is tidier because previously if the ClutterText did have
focus then it would be given all of the unhandled events twice.

Fixes bug 4137