[table] re-write the algorithm to calculate row heights
This new algorithm is more straight forward and also takes into account
the minimum and preferred size of children that span multiple rows.

This is essentially the same algorithm as that used to calculate column
widths, but uses the calculated column widths to help get the preferred
width of the children.