[clipboard] warn if a clipboard request is received but no text is available
In some circumstances, MxClipboard receives a clipboard request even
though it has not taken ownership of the clipboard selection. This
appears to happen when a GtkClipboard is used by the application to copy
text to the clipboard, so MxClipboard receives the selection request sent
for the GtkClipboard.

Therefore, to prevent a segmentation fault if clipboard_text has not been
set, we print a warning and return.

Fixes http://bugzilla.moblin.org/show_bug.cgi?id=5015