Simplify the Stylable interface
Remove the accessor functions from the Stylable interface and use
object properties instead. Also, remove the properties from the interface
that are not used or not implemented. The implementation of properties
that should be implemented the same way has also been moved out of the
interface requirements.

This makes it much simpler to implement the stylable interface from
a subclass of ClutterActor.