[styling] support border-image properties
To support the border-image property, the source filename for each property
needs to be returned in combination with the value so that url fields with
relative paths can be resolved.

Also, a new function parse the border image string, complete the relative
path and transform the information into a MxBorderImage.