[popup] paint and pick after the stage paint and pick have run
Allow the popup to be parented on the actor it is associated with and run
the pain and pick routines after the stage paint and pick have run to
allow the popup to appear above all other actors.

Add a test case to test-combo-box to demonstrate this and alter combo box
for the new behaviour.