[widget] move the long-press signal from MxButton to MxWidget
The long-press signal is useful to more than just button widgets, so expose
it from the MxWidget abstract class.

This patch also adds utility functions for MxWidget subclasses to emit the
long-press query, which depending on the return value, may or may not start
a long-press timer. The long-press cancel signal is also available in the
same way.