[application] Add actions to applications
Allow the adding, removing and invokation of MxActions to MxApplication.
This doesn't do the necessary DBus bits yet and MxAction probably needs
extending to make this more flexible.

Also, for multiple-instance applications, have the first register on the
bus and subsequent applications join the queue (instead of just not
registering).