Add MxFocusManager object and MxFocusable interface
MxFocusManager is an object that handles managing keyboard focus on a
ClutterStage. It intercepts various keyboard events and asks the current
focused actors to move or accept focus as required.

MxFocusable provides an interface for actors to implement to allow the
focus manager to move or apply the focus between widgets.