[table] implement MxFocusable
Implement MxFocusable on MxTable so that children of MxTable can be
navigated using MxFocusManager.

Although table cannot be focused directly, it can implement the moving of
focus withing its children by implementing the move_focus virtual function.

If the table is asked directly to accept focus it will find the first child
willing to accept focus.