[path-bar] Different styling for the last crumb
The last crumb in the bar is styled differently when the bar isn't
editable (so it doesn't end in an arrow pointing to nothing).

I'd have liked to do this by setting the style class, but this didn't
seem to work, so it's done using the actor name, similarly to the
scroll-bar.