[icon-theme] First run at an icon-theme object
Implement an object that lets you lookup icons from the default or a
specified theme. Currently, it will get the theme from an envvar
'MX_ICON_THEME', or it'll fallback to 'moblin'.

Also fix a possible leak in MxTextureCache and add the ability for it to
return cogl texture handles as well as actors.