[window] Fix minimum window size
Due to some bad interaction with the get_preferred_width/height
functions, the window size couldn't be shrunk to less than the default
640x480 in some cases. Change how these functions work so that they
report the user-set size as natural width/height.

This fixes this issue, though may cause issues when window managers try
to resize the window. Clutter could do with some better API to not
require this (there are FIXMEs in the Clutter X11 backend for this).