mx-texture-cache: Use CoglSubTexture instead of MxSubtexture
Since Clutter 1.1.8 Cogl directly has support for sub textures so we
can avoid the need for MxSubtexture. This has the advantage that there
are no differences to applications between textures retrieved from a
pre-generated cache and those loaded from a file directly. For example
it should now be possible to use a pre-generated cache image in a
texture frame.

http://bugs.meego.com/show_bug.cgi?id=2869