[window] Clean up mouse event handling
Previously it was very easy to get a mouse button 'stuck' during
resizing, or to not be able to resize due to a button under the resize
handle and such.

The resize handle now takes precedence over everything and we use grabs
and some event checking to make sure that the button never gets stuck.

We also now check the particular device that initiates a move/drag and
only respond to mouse-button 1. This frees up the other mouse buttons
for either new actions in the future, or to be used by child widgets.