[toggle] Allocate correctly regardless of size
When allocated areas larger than its size, the handle position would be
calculated incorrectly. Now, when a background-image is used, the handle
and background will correctly appear centred in the widget, and when
using a border-image, it will appear scaled.

Also modify relevent tests to use toggles instead of buttons.