[application] Add automagic screen-mode switching
Mutter-moblin exports a set of properties on the WM window to say
whether it's in small-screen or big-screen mode. Use and listen to this
property so that windows in an application automatically switch mode.

Note that this currently applies to all windows in an application and
there's no way to switch it off, we will need to revisit this policy for
multi-window applications and the like.