[widget] Fix bugs with old-bg in map/unmap/dispose
The old background wasn't getting mapped or unmapped, causing odd
artifacts when changing the style of a widget that hadn't been mapped
yet (as shown in test-mx on the style test).

This should also fix the possibility of an invalid pointer access in the
old background fading animation.

Possibly fixes Moblin bug #9364