[droppable] Don't hide the draggable before picking
Hiding the draggable does the wrong thing when the drabbable is left
inside its original container and that container does not have the
CLUTTER_ACTOR_NO_LAYOUT flag (which is the case for most containers
really, only containers that defer their layouting to ClutterFixedLayout
have that flag for now).

More specifically, when hiding a draggable, it will cause a relayout,
and then the pick buffer will be painted with this layout instead of the
one used to paint the color buffer which leads to some obvious mismatch
for the over-in/over-out/drop signals.

Instead of hiding the draggable, we set it non-reactive so it's not
picked.

Fixes: http://bugzilla.moblin.org/show_bug.cgi?id=9978