Add clutter-text properties to MxEntry and MxLabel
Based on patch from: Owen W. Taylor <otaylor@fishsoup.net>

Add clutter-text properties to allow getting access to the underlying
ClutterText actor. This corresponds to the get_clutter_text() methods.

The PROP_LABEL and PROP_ENTRY enum values are renamed to PROP_TEXT to
match the names of the properties that they correspond to, and the
properties of MxEntry are reordered into alphabetical order.

Based on a patch from Colin Walters
https://bugzilla.gnome.org/show_bug.cgi?id=591245

Fixes: http://bugzilla.moblin.org/show_bug.cgi?id=6313