window: overload ClutterContainer::add_actor() to support MxDraggables
For now, setting a new child on the window, either by the
ClutterContainer interface or by mx_window_set_child() replaces the
single child MxWindow can have, so we can't add the dragged actor.

Let's provide a way to use clutter_container_add_actor() to add a
MxDraggable. When adding a new actor we first check if it's a Draggable
(or if it's a clone, that its source is a Draggable) and update
priv->draggable with that. Then it's another internal actor we paint
after the child.