offscreen: Add a cogl-program property
The ClutterShader API is pretty woeful, unfortunately, and when combined
with MxOffscreen, you tend to get incorrect results due to it being
applied to all its children.

Add a cogl-program property to set a Cogl program that will get used
when rendering the texture only, and a test demonstrating how to use it
(vs. using ClutterShader).