Remove 'allocate-hidden' properties
We never allocate hidden. If a user wishes to allocate hidden children,
instead of hiding them, they can set their opacity to 0, or block/hide
their drawing in some other way.

This removes the 'allocate-hidden' property from MxGrid, and there was a
left-over unused 'allocate-hidden' property in MxTableChild, which this
also removes.