window: Sub-class GObject instead of ClutterStage
Sub-classing ClutterStage left us with various problems. For example, by
overriding the ClutterContainer interface implementation, ClutterGroup
API would fail to work on an MxWindow. Also, we had to special-case drag
actors so that they could be added to the stage and drawn over the
primary child of the window.

This changes MxWindow to be a sub-class of GObject and have a
'clutter-stage' property (that can be set on construct).

This has the advantage of allowing normal stage behaviour and reducing
the amount of code in MxWindow. The API is the same, except there is now
an extra function call, mx_window_new_with_clutter_stage.