item-view: Add 'factory' property and fix setter
Add a 'factory' property to mirror the set_factory function. There was
also no reason for there not to be a get function, so added this too.
The set factory function was missing a type-check and also would crash
when setting a NULL factory (which is valid), so also fix this.