path-bar: Add an 'entry' property for the MxEntry
MxPathBar had a get_entry function, but no corresponding readable
property. Add this property and also fix test-path-bar.js and alter it
to use this property.

Note that I'm not adding a 'text' property, as the get/set_text
functions are just a convenience wrapper to save on having to retrieve
the entry and setting it there.