notebook: Make API actor-oriented
Change the API so that it's actor-oriented instead of number-oriented.
The 'page' property is now 'current-page' and is an actor. There are no
page numbers, though a user is free to create their own number
associations.

The animation is also a proper cross-fade now, though this looks odd on
actors that don't have opaque backgrounds.

Also update the JavaScript test.