tooltips: Only show tooltips after a short delay
Most toolkits only display tooltips if the mouse is held steady on a
widget for a certain period of time. This patch implements that for Mx
as well.

Instead of displaying tooltips on an enter event the tooltip is now
displayed on mouse motion events. Whenever a motion event is received
a timeout is installed to display the tooltip after 500ms. If another
motion event is received then the timeout is restarted. If the mouse
leaves the widget then the timeout is abandoned. That way the tooltip
will only be displayed if the mouse is held still over a widget for
500ms.

Once a tooltip is displayed then it enters 'browse mode'. In browse
mode the tooltip is shown immediately on an enter event with no
delay. This is so a user can hover over a series of icons in a toolbar
without having to wait for the timeout to find out what each one
does. Browse mode is left 500ms after a tooltip is hidden.

The implementation is based on that in the GTK+ source code. However
in GTK+ both of the timeouts are customizable but here they are
hard-coded. It might be a good idea to use the setting directly from
GTK+.

http://bugzilla.clutter-project.org/show_bug.cgi?id=2179