path-bar: Don't ever shrink the background
Previously, we'd allow the background to be shrunk to below the texture
height (but its preferred height would always be at least the backgroud
height). However, this produces nasty-looking results when a path-bar is
put into a BoxLayout and something else in the box-layout doesn't get
its preferred height (a common situation if you pack a scrollable widget
into a box-layout).

Avoid this by expanding the minimum height to the background height as
well.