button: Ensure the label style is updated on creation
If an MxButton is mapped without a label set, it will contain no
MxLabel. If the text is then set, it will be created, but the style will
be wrong as it hasn't had a chance to set the ClutterText attributes.

Fix this by checking if we're mapped when we create the label and
calling mx_stylable_apply_clutter_text_attributes().

Fixes MeeGo bug #1570