offscreen: Add accumulation buffer capability
Add two properties, 'accumulation-enabled' and 'accumulation-material',
that controls a new capability of MxOffscreen: Acting as an accumulation
buffer.

When 'accumulation-enabled' is set to TRUE, MxOffscreen creates an extra
buffer that it uses as an accumulation buffer. The default settings have
it blend the primary buffer with the last frame at 50% opacity.

Other effects can be achieved by retrieving the accumulation buffer
material ('accumulation-material') and changing the blend string.

Addresses MeeGo Bug #2490