offscreen: Add 'redirect-enabled' property
Add a property, 'redirect-enabled', that lets you bypass
offscreen-redirection and draw the child directly instead. This is useful
if you have an effect that you want to turn on and off without having to
reparent an actor outside of the MxOffscreen.

Based on patch from Chris Lord <chris@linux.intel.com>

Fixes MeeGo Bug #1852