mx-settings: prevent possible recursion during the creation of MxSettings
During the construction of MxSettings, the default values are changed from
reading the xsettings values. Therefore, the "style-changed" signal may
be emitted. If a handler on this signal requires MxSettings, recurusion can
occur because the singleton has not yet been saved. This can be avoided by
making sure the xsettings client is initialised only after the saving of
the singleton object. This is acheived by implementing the singleton in
the object's constructor virtual function, and initialising the xsettings
client in the constructed signal.