finger-scroll: Don't use grabs and make event capturing optional
Disable motion events and attach to captured events on the stage, instead
of using grabs (this is more reliable and plays friendlier with other
actors).

By default, use button-press-event instead of captured-event to initiate
the scrolling - captured-event can optionally be turned back on.

Also make the mouse button that initiates the scrolling configurable,
instead of hard-coding mouse-button 1.