winsys: Allow Mx to compile without X
Introduce "window systems". For now 'none' and 'x11' which allows to
compile Mx for X11 (the default) or for no specific window system (for
instance if we are compiling against a eglnative version of Clutter.

For now, it's only a few #ifdef's but it might need to be extended
further if one has to write a Windows backend.