window: Remove the base class and add MxNativeWindow
Instead of having an abstract base class that has structs the size of
a GObject and implementing MxWindow in the backend, implement a base
MxWindow and have it create an MxNativeWindow object (implemented in the
backend) to do backend-specific tasks.

This is the same way Clutter does it with ClutterStage and
ClutterStageWindow.