window: Do rotations with child transformations
Using cogl transformations on pre-paint/pre-pick broke functions like
clutter_actor_transform_stage_point, as there's no way for those functions
to know that an arbitrary transformation has been applied by a signal
handler.

Instead of doing it this way, set the position/rotation of the children
of the window. This also now allows you to add actors that float above
the window, to which rotation doesn't apply. The resize-grip is one such
actor, for example.