stack: Add initial implementation
Add implementation of new widget, MxStack. This is an actor that allows
you to stack multiple widgets on top of each other, with child properties
that allow for alignment and fill.

This is similar to ClutterBox with the Bin layout manager, but MxStack will
not allocate its children a larger size than it is allocated itself.