Use new key symbols (with fallback)
Clutter ≥ 1.3.13 has deprecated the old CLUTTER_<keysym> defines and
switched to to be CLUTTER_KEY_<keysym>, to avoid eventual collisions.

Since Mx builds agains Clutter 1.2, we need to add a version check and a
couple of compatibility defines, if we still want to compile with
CLUTTER_DISABLE_DEPRECATED turned on.

http://bugzilla.clutter-project.org/show_bug.cgi?id=2310