stack: Allocate non-filled children correctly
If an actor supported width-for-height or height-for-width, setting it not
to fill in a stack would not work correctly. It may be that we want to fix
this in mx_allocate_align_fill, as it may affect other containers that use
this function, but for now, compensate for this in the allocate function
of MxStack.