style: Add a style cache
Instead of just storing the last gotten style matches and invalidating it
as soon as the style changes, store a cache of the last 6 applied styles
against a unique string representing the matchable properties of the
stylable object.

Invalidate this cache on map, parent-set, on name-change, when the style
is set, and when pseudo-class or style-class change. This significantly
cuts down the amount of necessary look-ups (at the cost of a small
amount of memory).