mx-widget: Set the hover state also when the pointer enters a child
The hover pseudo class on an actor is now also set on a widget
whenever the pointer enters any children of the actor. This is the
same behaviour as implemented in HTML. The widget tracks this by
ignoring leave events when the pointer is leaving to a child. The
actor should still get the final leave event when the pointer leaves
the child because it should bubble up to the parent.

Previously the enter and leave events would be ignored when the actor
is marked as disabled. I've removed this because it would break the
case where the pointer moves to a disabled child actor and then leaves
both actors without first passing through the parent. I think the code
isn't necessary anyway because mx_widget_get_pseudo_class forcibly
overrides the class when the widget is disabled so it shouldn't do any
harm to set the hover state.

http://bugzilla.clutter-project.org/show_bug.cgi?id=2357