Don't compile with *_DISABLE_DEPRECATED
The G_DISABLE_DEPRECATED and CLUTTER_DISABLE_DEPRECATED flags are only
really intended for debugging and it doesn't make much sense to set
them in all builds. Even if Mx is not currently using any deprecated
features the defines may make the build fail with future versions of
Clutter or glib that would otherwise be backwards compatible.