kinetic-scroll-view: Ensure child has adjustments
When adding a child to an MxKineticScrollView, call
mx_scrollable_get_adjustments() to ensure that the adjustments have been
created before allocation has been reached. This fixes the case where
adding an MxBoxLayout to an MxKineticScrollView can result in unexpected
behaviour, as its adjustments are initialised in its own allocate
function. This is analogous to the behaviour of MxScrollView.

Ideally, MxBoxLayout should initialise its adjustments immediately using
the current allocation box, but there are cases where this may cause
allocation cycles.