icon: Add a new style property '-mx-icon-suffix' + bug-fixes
Add a new style property, '-mx-icon-suffix', that allows you to style an
extra suffix to append to the icon name. This is handy if you have multiple
states of a single icon that depend on stylable states.

This also fixes programmatically setting icon name/size not overriding
the styled name/size, and the memory leak of the icon name on destruction.