css: check more precisely for pseudo-class matches
The pseudo-class list from the selector needs to be a subset of the
pseudo-class list from the node for a correct match. The inclusion of an
element in the list cannot be determined using strstr, since this may
incorrectly match on an element that includes the string of another
element.