image: add a 'clear' function to replace the current image with a blank one
mx_image_clear will replace the current image with a blank, transparent
image. MxImage will also now be initialised with the blank texture so that
the first cross fade will be between a transparent image and the new image,
rather than cross fading with an invalid texture handle as the old image.