item/list-view: Improve the freeze/thaw behaviour
Use an internal flag to determine whether the view is frozen, rather than
blocking the signal handlers. This means that if the list or item view is
frozen and a model is set, the view will still remain frozen until thaw
is called, even though new signal handlers have been connected.

http://bugs.meego.com/show_bug.cgi?id=1195