mx-window-x11: Fix mx_window_x11_set_wm_hints for 64-bit archs
For some reason XChangeProperty always takes an array of longs when
the format is 32 even on architectures where sizeof(long) != 32. The
RGBA data we get from Cogl is packed into 32-bit values so on 64-bit
architectures we need to expand these values to longs. This was
causing a crash when using a custom icon on 64-bit because libX11
would try to read past the end of the data array.

http://bugzilla.clutter-project.org/show_bug.cgi?id=2533