build: Use the introspection's Makefile rules
The GObject Introspection projects ships with a nice Makefile template
that you can include into your own project's autotooling to get nicer
and standardized rules. This requires a bit of fiddling in order to have
versioned targets in our own Makefile.am, but it's trivial enough.

In this case, we copied the same setup Clutter uses.