build: Fix behaviour of --disable-gtk-widgets
This was caused by a mistake in the use of AC_ARG_ENABLE. The first parameter
is used when the option is used in any form, either --enable or --disable. In
this case the bug was that when any mode was used the enable value was
overriden to no meaning that --enable-gtk-widgets would actually disable gtk
widget building.

The correct solution here was to set the default value in the second case of
AC_ARG_ENABLE (not specified) and then use the actual enable value to decide
whether to enable or not.

Fixes: http://bugzilla.clutter-project.org/show_bug.cgi?id=2552