image: Fix some concurrency issues with asynchronous loading
Always allow the idle handler to free the async data structure, check that
it completed before setting the image, check that the buffer is valid
before trying to write it in the pixbuf loader and make sure loads are
cancelled properly before setting the image.

This should fix some possible warnings, and the possibility of an
asynchronous image load completing after a synchronous image load has been
done.