texture-cache: Add the ability to store meta-textures
Add new functions that allow you to store associated texture data with
URIs.

This is handy if you have a small modification that you need to make to the
actual texture data for a particular use, but you still want this modified
texture to be cached and reused between multiple actors.