table: Fix asking for width/height without a for_height/width failing
If you asked a table for its preferred width or height without also
specifying a for-width/height, it would always return just the minimum
size.

Fix this by assigning the final-size of the row/column to the preferred
size by default (this will later be overwritten if a for-width/height is
specified).