kinetic-scroll-view: Fix non-capture mode
When not using captured events, motion events were still captured before
the scroll threshold was exceeded, making it a bit less useful. Alter the
event-handling process so that in non-capture mode, the motion-event
signal is used instead of the captured-event signal before the scroll
threshold is exceeded.

This allows embedding horizontally scrolling kinetic scroll-views in
vertically scrolling ones (and vice-versa).