mx-offscreen: Handle redirecting a ClutterStage specially
When redirecting a ClutterStage there's no need to calculate our own
modelview matrix because the one for the stage will already be set up
render to a buffer of the same size. This gives the best chance of
rendering exactly the same geometry as would be rendered to the stage
even if Clutter changes how it calculates its modelview matrix. This
is important when the offscreen is used to temporarily apply an effect
to the stage such as from MxDialog.

This also makes it avoid calling cogl_clear for stages because the
paint implementation for the stage also does a clear so it's a
waste. Having the clear there also allows Mx to take advantage of
Clutter's optimisation to avoid clearing when the
CLUTTER_STAGE_NO_CLEAR_ON_PAINT hint is set.

http://bugzilla.clutter-project.org/show_bug.cgi?id=2658